[ADD] add some css and code optimizations
