+++ client/src/components/Contact.css
@@ -1,0 +1,3 @@
+.fa-trash {
+   color: white; 
+}
