+++ client/src/components/ContactList.css
@@ -1,0 +1,4 @@
+.nocontact {
+    text-align: center;
+    color: lightcoral;
+}
