+++ client/src/components/Header.js
@@ -13,14 +13,13 @@ class Header extends React.Component {
     }
 
     logout = () => {
-        this.props.logout();
-        this.props.history.push('/');
+        this.props.logout(() => {this.props.history.push('/');});
     }
 
     render() {
         return (
             <Navbar bg="light" expand="lg">
-                <Navbar.Brand href="#home" style={{ color: 'red' }}>Contact Manager</Navbar.Brand>
+                <Navbar.Brand style={{ color: 'red' }} href="/home/contactlist">Contact Manager</Navbar.Brand>
                 <Navbar.Toggle aria-controls="basic-navbar-nav" />
                 <Navbar.Collapse id="basic-navbar-nav">
                     <Nav className="ml-auto">
