+++ client/src/components/Login.css
@@ -1,0 +1,8 @@
+.center {
+    width:300px;
+    height:200px;
+    position:absolute;
+    left:50%;
+    top:50%;
+    margin:-100px 0 0 -150px;
+}
